% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{county_codes}
\alias{county_codes}
\title{North Carolina County Codes}
\format{a dataframe with 100 rows and 2 columns:
\describe{
  \item{county_number}{North Carolina County Number}
  \item{county_name}{The name of the county}
   ...
}}
\usage{
county_codes
}
\description{
North Carolina County Codes
}
\keyword{datasets}
