% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voters.R
\name{get_voters}
\alias{get_voters}
\title{Get NC voter Data}
\usage{
get_voters(file_type, county, local_version = FALSE,
  out_location = NULL)
}
\arguments{
\item{file_type}{the file you would like to retrieve}

\item{county}{the county you would like to retrieve}

\item{local_version}{boolean to write a local SQLlite instance. Default of
\code{FALSE}}
}
\description{
Get NC voter Data
}
\examples{
\dontrun{

dat <- get_voters(file_type = "ncvhis", county = "forsyth")
dat <- get_voters("ncvhis", 67)
}
}
