% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvoter.R
\docType{package}
\name{ncvoter-package}
\alias{ncvoter}
\alias{ncvoter-package}
\title{ncvoters}
\description{
A set of functions to download NC Voter Rolls and History
}
\author{
\strong{Maintainer}: Michael DeWitt \email{me.dewitt.jr@gmail.com} (0000-0001-8940-1967)

}
